% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{contaminate_sample}
\alias{contaminate_sample}
\title{Contaminate a sample on different quantile based cutpoints}
\usage{
contaminate_sample(x, percent = 0.04, location = "bq")
}
\arguments{
\item{x}{Input sample}

\item{percent}{Percentage of contamination}

\item{location}{Options: "bq" - Both quantiles, "low" - Inferior quantile,
"high" - Superior quantile, "bm" - Both minimum and maximum, 
"min" - Minimum, "max" - Maximum}
}
\value{
Contaminated sample
}
\description{
Contaminate a sample on different quantile based cutpoints
}

