\name{KB36_t}
\alias{KB36_t}
\docType{data}

\title{Data on two 36-items test forms}

\description{
The data set contains item parameters estimates following a 3PL model for two 36-items tests forms, 
rescaled using mean-sigma method's A and B using all common items except item 27. This data has been described and analized by Kolen and Brennan (2004), Table 6.8.}

\usage{data(KB36_t)}

\format{A dataframe where each column represent item parameter estimates of forms \code{X} and \code{Y}, with their respective p-values.
}

\references{

	Kolen, M., and Brennan, R. (2004). \emph{Test Equating, Scaling and Linking}. 
	New York, NY: Springer-Verlag.
}

\seealso{KB36}

\examples{
data(KB36_t)
}

\keyword{datasets}
