\name{Math20EG}
\alias{Math20EG}
\docType{data}

\title{Scores on two 20-items mathematics tests.
}

\description{
The data set contains raw sample frequencies of number-right scores for two parallel 
20-items mathematics tests given to two samples from a national population of 
examinees. This data has been described and analized by Holland and Thayer (1989); 
Von Davier et al, (2004) (see also Von Davier, 2011 where other applications using 
these data set are shown).
}

\usage{data(Math20EG)}

\format{A 21x2 matrix containing raw sample frequencies (raws) for two parallel tests (columns) 
}

\references{
	Holland, P. and Thayer, D. (1989). The kernel method of equating score distributions. 
	(Technical Report No 89-84). Princeton, NJ: Educational Testing Service.

	Von Davier, A., Holland, P., and Thayer, D. (2004). \emph{The Kernel Method of Test Equating}. 
	New York, NY: Springer-Verlag.  
}
\examples{
data(Math20EG)
## maybe str(Math20EG) ; ...
}
\keyword{datasets}
