\name{Math20SG}
\alias{Math20SG}
\docType{data}
\title{Bivariate score frequencies on two 20-items mathematics tests.
}
\description{The data set contains the bivariate sample frequencies of number-right scores 
for two parallel 20-items mathematics tests given to a sample from a national population of 
examinees. This data has been described and analized by Holland and Thayer (1989); 
Von Davier et al, (2004).
}

\usage{data(Math20SG)}

\format{A 21x21 matrix containing the bivariate sample frequencies for \eqn{X}{X} (raws) and \eqn{Y}{Y} 
(columns) 
}

\references{
	Holland, P. and Thayer, D. (1989). The kernel method of equating score distributions. 
	(Technical Report No 89-84). Princeton, NJ: Educational Testing Service.

	Von Davier, A., Holland, P., and Thayer, D. (2004). \emph{The Kernel Method of Test Equating}. 
	New York, NY: Springer-Verlag.  
}
\examples{
data(Math20SG)
## maybe str(Math20SG) ; ...
}
\keyword{datasets}
