% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuncs.R
\name{pasted_table_to_df}
\alias{pasted_table_to_df}
\title{Transform a table from String to Data Frame.}
\usage{
pasted_table_to_df(string, n_cols, sep = " ", header = TRUE)
}
\arguments{
\item{string}{Input string}

\item{n_cols}{Number of columns in the table}

\item{sep}{Separator character}

\item{header}{Have a header?}
}
\value{
data.frame
}
\description{
The usual way to compare results is using tables from books. This method
accept as input a pasted String, usually copied from a LaTeX table, and
create a Data Frame. Assumes the input is separated evenly by a character,
default is an empty space.
}

