% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuncs.R
\name{rowBlockSum}
\alias{rowBlockSum}
\title{Take a matrix and sum blocks of rows}
\usage{
rowBlockSum(mat, blocksize, w = NULL)
}
\arguments{
\item{mat}{Input matrix}

\item{blocksize}{Size of the row blocks}

\item{w}{(Optional) Vector for weighted sum}
}
\value{
Matrix
}
\description{
The original data set contains very long column headers. This function
does a keyword search over the headers to find those column headers that
match a particular keyword, e.g., mean, median, etc.
}

