\name{SEPA}
\alias{SEPA}
\docType{data}
\title{A sample of observed score values for two different forms of the SEPA test. 
}
\description{The data set is from a private national evaluation system called SEPA. It
contains two test forms X and Y both composed of 50 items.
The SEPA data is a list containing two samples with 1,458 test takers who
took test form X and 2,619 test takers who took test form Y. 
}
\usage{data(SEPA)}
\format{
A list with elements containing the observed scores in test forms X and Y.
}

\references{
	Gonzalez, J. and Wiberg, M. (2017). \emph{Applying test equating methods using R}. Springer. 
}
\examples{
data(SEPA)
## maybe str(SEPA) ; ...
}
\keyword{datasets}