% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_unimodal}
\alias{sim_unimodal}
\title{Simulate different unimodal distributions.}
\usage{
sim_unimodal(n, x_mean, x_var, N_item, seed = NULL, name = NULL)
}
\arguments{
\item{n}{Size of the simulated distribution}

\item{x_mean}{Mean}

\item{x_var}{Variance}

\item{N_item}{Number of items simulated}

\item{seed}{(Optional) Set a random seed}

\item{name}{(Optional) Sample based on the names given by 
Keats & Lord (1962). 
Options are: "GANA", "MACAA", "TQS8", "WM8", "WMI".}
}
\value{
Simulated values
}
\description{
Simulate different unimodal, skewed distributions based
on different mean and variance parameters.
}
\details{
All the inner working of this procedure can be seen 
in Keats & Lord (1962).
}

