% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tLDBPPdensity.R
\name{tLDBPPdensity}
\alias{tLDBPPdensity}
\alias{tLDBPPdensity.default}
\title{tLDBPPdensity.}
\usage{
tLDBPPdensity(formula, xpred, prior, mcmc, state, status, ngrid = 100,
  grid = NULL, compute.band = FALSE, type.band = "PD",
  data = sys.frame(sys.parent()), na.action = na.fail, work.dir = NULL)
}
\arguments{
\item{formula}{asd}

\item{xpred}{asd}

\item{prior}{asd}

\item{mcmc}{asd}

\item{state}{asd}

\item{status}{asd}

\item{ngrid}{=100,}

\item{grid}{=NULL,}

\item{compute.band}{=FALSE,}

\item{type.band}{="PD",}

\item{data}{=sys.frame(sys.parent()),}

\item{na.action}{=na.fail,}

\item{work.dir}{=NULL}
}
\value{
Simulated values
}
\description{
tLDBPPdensity.
}

