% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP.eq.R
\name{BNP.eq.predict}
\alias{BNP.eq.predict}
\title{Prediction step for Bayesian non-parametric model for test equating}
\usage{
BNP.eq.predict(model, from = NULL, into = NULL, alpha = 0.05)
}
\arguments{
\item{model}{A 'BNP.eq' object.}

\item{from}{Numeric. A vector of indices indicating from which patterns equating should be performed.
The covariates involved are integrated out.}

\item{into}{Numeric. A vector of indices indicating into which patterns equating should be performed.
The covariates involved are integrated out.}

\item{alpha}{Numeric. Significance for credible bands.}
}
\value{
A 'BNP.eq.predict' object, which is list containing the following items:

pdf A list of PDF's.

cdf A list of CDF's.

equ Numeric. Equated values.

grid Numeric. Grid used to evaluate pdf's and cdf's.
}
\description{
The
}
\details{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
Vivamus finibus vitae eros quis dictum. Donec lacus risus, facilisis quis 
tincidunt et, tincidunt sed mi. Nullam ullamcorper eros est, sed 
fringilla metus volutpat eu. Etiam ornare nulla id lorem posuere, eu vehicula 
urna vestibulum. Quisque luctus, diam ac mattis faucibus, leo felis tincidunt urna, eu 
tempus massa neque nec nibh. Aliquam erat volutpat. Fusce tempor mattis enim quis pretium. 
Aliquam volutpat luctus felis, nec fringilla enim tincidunt sed. 
Nam nec leo quis erat lobortis vulputate ac at neque
}
\author{
Daniel Leon A. \email{dnacuna@mat.uc.cl}, Felipe Barrientos \email{afb26@stat.duke.edu}.
}
\references{
asdasd
}
\keyword{BNP}
\keyword{Bayesian}
\keyword{equating}
\keyword{equating,}
\keyword{non-parametrics,}

