\name{SNSequate-package}
\alias{SNSequate-package}
\alias{SNSequate}
\docType{package}
\title{Standard and Nonstandard Statistical Models and Methods for Test
Equating
}

\description{The package contains functions to perform various models and methods for test equating. It currently
implements the Kernel Method of Equating using the "traditional" Gaussian kernel and also
nonstandard alternative kernels such as the logistic and uniform. More kernels will be implemented
in the near future as well as other nonstandard methods of equating. Functions to obtain both standard error of 
equating (SEE) and standard error of equating difference between two equating functions (SEED) are also implmented.
}

\details{
\tabular{ll}{
Package: \tab SNSequate\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-30\cr
License: \tab GPL (>= 2)\cr
}
}

\author{Jorge Gonzalez Burgos

Maintainer: Jorge Gonzalez Burgos <jgonzale@mat.puc.cl>

}

\references{Von Davier et. al (2004): The Kernel Method of Test Equating
}


\keyword{ package }

